lib: relocate prob from phase2/ to lib/
The CDF / inverse-sampling kernel is a generic numerical
helper used by the qdrift and cmpsit samplers; it doesn't
depend on any phase2 type and has its own t-prob unit
test.  Move it next to its peers in lib/:

  phase2/prob.c          -> lib/prob.c
  include/phase2/prob.h  -> include/prob.h

The flat header location matches the lib convention
(include/log.h, include/xoshiro256ss.h, etc.).  Callers'
includes update from "phase2/prob.h" to "prob.h"; the top-
level phase2.h re-export and the per-subsys Makefile SRCS
lists move accordingly.  Pure mechanical rearrangement;
the implementation is untouched.
